Safe Zone at start
Created a boolean named firstClick, changed NewGame() has GenerateMines/Numbers commented out and set firstClick to true, changed GenerateMines to grab mouse world position, change it to board position, and check to see if the random generation wasn't on that spot or any of the adjacent spots. Changed Update() so the else if case for 0 checks if firstClick is true, if so call the Generates, set to false and reveal, if not then just reveal.